+++ vemproifood-dish-list-service/src/main/java/br/com/ifood/vemproifood/dishlist/model/AddressComponent.java
@@ -1,0 +1,33 @@
+package br.com.ifood.vemproifood.dishlist.model;
+
+import java.util.List;
+
+public class AddressComponent {
+	private String long_name;
+	private String short_name;
+	private List<String> types;
+
+	public String getLong_name() {
+		return long_name;
+	}
+
+	public void setLong_name(String long_name) {
+		this.long_name = long_name;
+	}
+
+	public String getShort_name() {
+		return short_name;
+	}
+
+	public void setShort_name(String short_name) {
+		this.short_name = short_name;
+	}
+
+	public List<String> getTypes() {
+		return types;
+	}
+
+	public void setTypes(List<String> types) {
+		this.types = types;
+	}
+}
